Move BPE tests to their own file
