+++ vectorizers/tests/__init__.py
@@ -1,0 +1,8 @@
+raw_string_data = [
+    "asdfj;afoosdaflksapokwerfoobarpokwersdfsadfsadfnbkajyfoopokwer",
+    "pokfoo;ohnASDbarfoobarpoksdf sgn;asregtjpoksdfpokpokwer",
+    "werqweoijsdcasdfpoktrfoobarpokqwernasdfasdpokpokpok",
+    "pokwerpokwqerpokwersadfpokqwepokwerpokpok",
+    "foobarfoofooasdfsdfgasdffoobarbazcabfoobarbarbazfoobaz",
+    "pokfoopokbarpokwerpokbazgfniusnvbgasgbabgsadfjnkr[pko",
+]
